[storage] StateKey::get_shard_id should return usize
Returning `usize` seems to eliminate unnecessary type conversion in a bunch of
places and simplify the code a bit.

+++ types/src/nibble/nibble_path/mod.rs
@@ -221,9 +221,9 @@ impl NibblePath {
     }
 
     // Returns the shard_id of the NibblePath, or None if it is root.
-    pub fn get_shard_id(&self) -> Option<u8> {
+    pub fn get_shard_id(&self) -> Option<usize> {
         if self.num_nibbles() > 0 {
-            Some(u8::from(self.get_nibble(0)))
+            Some(usize::from(self.get_nibble(0)))
         } else {
             None
         }
